+++ crypto/evp/p_rsa_asn1.c
@@ -64,6 +64,7 @@
 
 #include "../fipsmodule/rsa/internal.h"
 #include "internal.h"
+#include "../rsa_extra/rsassa_pss.h"
 
 
 static int rsa_pub_encode(CBB *out, const EVP_PKEY *key) {
