+++ crypto/fipsmodule/rsa/rsa.c
@@ -77,6 +77,7 @@
 #include "../../internal.h"
 #include "../../evp/internal.h"
 #include "internal.h"
+#include "../../rsa_extra/rsassa_pss.h"
 
 
 // RSA_R_BLOCK_TYPE_IS_NOT_02 is part of the legacy SSLv23 padding scheme.
@@ -119,74 +120,6 @@ RSA *RSA_new_method(const ENGINE *engine) {
   }
 
   return rsa;
-}
-
-RSA_INTEGER *RSA_INTEGER_new(void) {
-  RSA_INTEGER *ret = OPENSSL_malloc(sizeof(RSA_INTEGER));
-  if (ret == NULL) {
-    OPENSSL_PUT_ERROR(RSA, ERR_R_MALLOC_FAILURE);
-    return NULL;
-  }
-  OPENSSL_memset(ret, 0, sizeof(RSA_INTEGER));
-  return ret;
-}
-
-RSA_ALGOR_IDENTIFIER *RSA_ALGOR_IDENTIFIER_new(void) {
-  RSA_ALGOR_IDENTIFIER *ret = OPENSSL_malloc(sizeof(RSA_ALGOR_IDENTIFIER));
-  if (ret == NULL) {
-    OPENSSL_PUT_ERROR(RSA, ERR_R_MALLOC_FAILURE);
-    return NULL;
-  }
-  OPENSSL_memset(ret, 0, sizeof(RSA_ALGOR_IDENTIFIER));
-  return ret;
-}
-
-RSA_MGA_IDENTIFIER *RSA_MGA_IDENTIFIER_new(void) {
-  RSA_MGA_IDENTIFIER *ret = OPENSSL_malloc(sizeof(RSA_MGA_IDENTIFIER));
-  if (ret == NULL) {
-    OPENSSL_PUT_ERROR(RSA, ERR_R_MALLOC_FAILURE);
-    return NULL;
-  }
-  OPENSSL_memset(ret, 0, sizeof(RSA_MGA_IDENTIFIER));
-  return ret;
-}
-
-RSASSA_PSS_PARAMS *RSASSA_PSS_PARAMS_new(void) {
-  RSASSA_PSS_PARAMS *ret = OPENSSL_malloc(sizeof(RSASSA_PSS_PARAMS));
-  if (ret == NULL) {
-    OPENSSL_PUT_ERROR(RSA, ERR_R_MALLOC_FAILURE);
-    return NULL;
-  }
-  OPENSSL_memset(ret, 0, sizeof(RSASSA_PSS_PARAMS));
-  return ret;
-}
-
-void RSA_INTEGER_free(RSA_INTEGER *ptr) {
-  OPENSSL_free(ptr);
-}
-
-void RSA_ALGOR_IDENTIFIER_free(RSA_ALGOR_IDENTIFIER *algor) {
-  OPENSSL_free(algor);
-}
-
-void RSA_MGA_IDENTIFIER_free(RSA_MGA_IDENTIFIER *mga) {
-  if (mga == NULL) {
-    return;
-  }
-  RSA_ALGOR_IDENTIFIER_free(mga->mask_gen);
-  RSA_ALGOR_IDENTIFIER_free(mga->one_way_hash);
-  OPENSSL_free(mga);
-}
-
-void RSASSA_PSS_PARAMS_free(RSASSA_PSS_PARAMS *params) {
-  if (params == NULL) {
-    return;
-  }
-  RSA_ALGOR_IDENTIFIER_free(params->hash_algor);
-  RSA_MGA_IDENTIFIER_free(params->mask_gen_algor);
-  RSA_INTEGER_free(params->salt_len);
-  RSA_INTEGER_free(params->trailer_field);
-  OPENSSL_free(params);
 }
 
 void RSA_free(RSA *rsa) {
