+++ include/openssl/base.h
@@ -412,10 +412,6 @@ typedef struct private_key_st X509_PKEY;
 typedef struct rand_meth_st RAND_METHOD;
 typedef struct rc4_key_st RC4_KEY;
 typedef struct rsa_meth_st RSA_METHOD;
-typedef struct rsa_integer_st RSA_INTEGER;
-typedef struct rsa_pss_supported_algor_st RSA_PSS_SUPPORTED_ALGOR;
-typedef struct rsa_algor_identifier_st RSA_ALGOR_IDENTIFIER;
-typedef struct rsa_mga_identifier_st RSA_MGA_IDENTIFIER;
 typedef struct rsassa_pss_params_st RSASSA_PSS_PARAMS;
 typedef struct rsa_st RSA;
 typedef struct sha256_state_st SHA256_CTX;
